Updates to original and alt control version
Found missing line in original macro. The initial parameters were being set to 0 and it was causing problems. Also updated macro_alt for taking tech_alt using new generation control function. This generation control function lets the battery supply power in the morning (while PV is on). Previously batteries could only discharge when PV was off, which meant regardless of BattSOC the propane generator turned on. This fixes that.